+++ x-uploader/test/x/uploader/media_request_test.rb
@@ -53,9 +53,9 @@ module X
     end
 
     def test_upload_rejects_missing_file_before_reading_it
-      error = assert_raises(RuntimeError) { Uploader::Media.upload("nope.jpg", client: @client, media_category: "tweet_image") }
+      error = assert_raises(Errno::ENOENT) { Uploader::Media.upload("nope.jpg", client: @client, media_category: "tweet_image") }
 
-      assert_equal "File not found: nope.jpg", error.message
+      assert_equal "No such file or directory - nope.jpg", error.message
     end
 
     private